# Update latencies, log track stats