--- a/go.mod
+++ b/go.mod
@@ -21,7 +21,7 @@ require (
 	github.com/livekit/mageutil v0.0.0-20230125210925-54e8a70427c1
 	github.com/livekit/protocol v1.5.8-0.20230601212100-a186ecb11a98
 	github.com/livekit/psrpc v0.3.0
-	github.com/livekit/server-sdk-go v1.0.11-0.20230603013036-a59560123fed
+	github.com/livekit/server-sdk-go v1.0.11-0.20230603013535-cd2ade940f03
 	github.com/pion/rtp v1.7.13
 	github.com/pion/webrtc/v3 v3.2.8
 	github.com/prometheus/client_golang v1.15.1