Raised web latency from 1.5s to 2s because we were still getting flvmux `backwards dts` warnings
Lowered sdk latency from 4s to 3s because the jitter buffer max latency is set to 2s (and works now)
Updated flvmux latency to match source latency